requirements.txt
torch>=1.1
pytorch-ignite
matplotlib
